+++ src/Exceptions/StuckJobsDetected.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Okipa\LaravelStuckJobsNotifier\Exceptions;
+
+use Exception;
+
+class StuckJobsDetected extends Exception
+{
+    //
+}