+++ src/Notification.php
@@ -38,7 +38,7 @@ class Notification extends IlluminateNotification
         return (new MailMessage)->error()->subject('[' . config('app.name') . ' - ' . config('app.env') . '] ⚠ '
             . $this->stuckJobsCount . ' stuck failed ' . Str::plural('job', $this->stuckJobsCount) . ' detected')
             ->line($this->stuckJobsCount . ' failed ' . Str::plural('job', $this->stuckJobsCount)
-                . ', stuck for at least ' . config('stuck-jobs-notifier.hoursLimit') . ' days, detected at '
+                . ', stuck for at least ' . config('stuck-jobs-notifier.hours_limit') . ' days, detected at '
                 . config('app.url') . '.')
             ->line('Please check your stuck jobs using the « php artisan queue:failed » command.');
     }
@@ -52,7 +52,7 @@ class Notification extends IlluminateNotification
     {
         return (new SlackMessage)->error()->content('⚠ `' . config('app.name') . ' - ' . config('app.env') . '` '
             . $this->stuckJobsCount . ' failed ' . Str::plural('job', $this->stuckJobsCount) . ', stuck for at least '
-            . config('failed-jobs-notifier.hoursLimit') . ' days, detected at ' . config('app.url') . '.');
+            . config('stuck-jobs-notifier.hours_limit') . ' days, detected at ' . config('app.url') . '.');
     }
 
     /**
@@ -66,7 +66,7 @@ class Notification extends IlluminateNotification
         return WebhookMessage::create()->data([
             'text' => '⚠ `' . config('app.name') . ' - ' . config('app.env') . '` ' . $this->stuckJobsCount
                 . ' failed ' . Str::plural('job', $this->stuckJobsCount) . ', stuck for at least '
-                . config('failed-jobs-notifier.hoursLimit') . ' days, detected at ' . config('app.url') . '.',
+                . config('stuck-jobs-notifier.hours_limit') . ' days, detected at ' . config('app.url') . '.',
         ])->header('Content-Type', 'application/json');
     }
 }